Consolidate Linear GraphQL client error handling.
Move query templating onto GraphqlClient, remove templating.py, and raise LinearActionError from API responses without per-call error_prefix strings.
